Embed `ToggleGroupControlButton` into `ToggleGroupControlOption`
