+++ packages/components/src/toggle-group-control/toggle-group-control-option/component.tsx
@@ -1,4 +1,10 @@
 /**
+ * External dependencies
+ */
+// eslint-disable-next-line no-restricted-imports
+import { Radio } from 'reakit';
+
+/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
@@ -12,9 +18,15 @@ import {
 	useContextSystem,
 	WordPressComponentProps,
 } from '../../ui/context';
-import ToggleGroupControlButton from './toggle-group-control-button';
-import type { ToggleGroupControlOptionProps } from '../types';
+import type {
+	ToggleGroupControlOptionProps,
+	ToggleGroupControlButtonProps,
+} from '../types';
 import { useToggleGroupControlContext } from '../context';
+import * as styles from './styles';
+import { useCx } from '../../utils/hooks';
+
+const { ButtonContentView, LabelPlaceholderView, LabelView } = styles;
 
 function ToggleGroupControlOption(
 	props: WordPressComponentProps< ToggleGroupControlOptionProps, 'input' >,
@@ -24,19 +36,49 @@ function ToggleGroupControlOption(
 	const id = useInstanceId(
 		ToggleGroupControlOption,
 		toggleGroupControlContext.baseId || 'toggle-group-control-option'
-	);
+	) as string;
 	const buttonProps = useContextSystem(
 		{ ...props, id },
 		'ToggleGroupControlOption'
 	);
+
+	const {
+		className,
+		isBlock = false,
+		label,
+		value,
+		...radioProps
+	}: ToggleGroupControlButtonProps = {
+		...toggleGroupControlContext,
+		...buttonProps,
+	};
+
+	const isActive = radioProps.state === value;
+	const cx = useCx();
+	const labelViewClasses = cx( isBlock && styles.labelBlock );
+	const classes = cx(
+		styles.buttonView,
+		className,
+		isActive && styles.buttonActive
+	);
+
 	return (
-		<ToggleGroupControlButton
-			forwardedRef={ forwardedRef }
-			{ ...{
-				...toggleGroupControlContext,
-				...buttonProps,
-			} }
-		/>
+		<LabelView className={ labelViewClasses } data-active={ isActive }>
+			<Radio
+				{ ...radioProps }
+				as="button"
+				aria-label={ radioProps[ 'aria-label' ] ?? label }
+				className={ classes }
+				data-value={ value }
+				ref={ forwardedRef }
+				value={ value }
+			>
+				<ButtonContentView>{ label }</ButtonContentView>
+				<LabelPlaceholderView aria-hidden>
+					{ label }
+				</LabelPlaceholderView>
+			</Radio>
+		</LabelView>
 	);
 }
 export default contextConnect(
